+++ packages/controller-utils/src/index.test.ts
@@ -6,10 +6,12 @@ describe('@metamask/controller-utils', () => {
       Array [
         "BrokenCircuitError",
         "CircuitState",
+        "ConstantBackoff",
         "DEFAULT_CIRCUIT_BREAK_DURATION",
         "DEFAULT_DEGRADED_THRESHOLD",
         "DEFAULT_MAX_CONSECUTIVE_FAILURES",
         "DEFAULT_MAX_RETRIES",
+        "ExponentialBackoff",
         "createServicePolicy",
         "handleAll",
         "handleWhen",