+++ packages/network-controller/src/create-network-client.ts
@@ -25,6 +25,8 @@ import {
 import type { JsonRpcMiddleware } from '@metamask/json-rpc-engine';
 import type { Hex, Json, JsonRpcParams } from '@metamask/utils';
 
+import type { NetworkControllerMessenger } from './NetworkController';
+import type { RpcServiceOptions } from './rpc-service/rpc-service';
 import { RpcServiceChain } from './rpc-service/rpc-service-chain';
 import type {
   BlockTracker,
@@ -51,20 +53,22 @@ export type NetworkClient = {
  *
  * @param args - The arguments.
  * @param args.configuration - The network configuration.
- * @param args.fetch - A function that can be used to make an HTTP request,
- * compatible with the Fetch API.
- * @param args.btoa - A function that can be used to convert a binary string
- * into base-64.
+ * @param args.getRpcServiceOptions - Factory for constructing RPC service
+ * options. See {@link NetworkControllerOptions.getRpcServiceOptions}.
+ * @param args.messenger - The network controller messenger.
+ * See {@link NetworkControllerOptions.getRpcServiceOptions}.
  * @returns The network client.
  */
 export function createNetworkClient({
   configuration,
-  fetch: givenFetch,
-  btoa: givenBtoa,
+  getRpcServiceOptions,
+  messenger,
 }: {
   configuration: NetworkClientConfiguration;
-  fetch: typeof fetch;
-  btoa: typeof btoa;
+  getRpcServiceOptions: (
+    rpcEndpointUrl: string,
+  ) => Omit<RpcServiceOptions, 'failoverService' | 'endpointUrl'>;
+  messenger: NetworkControllerMessenger;
 }): NetworkClient {
   const primaryEndpointUrl =
     configuration.type === NetworkClientType.Infura
@@ -74,12 +78,29 @@ export function createNetworkClient({
     primaryEndpointUrl,
     ...configuration.failoverRpcUrls,
   ];
-  const rpcService = new RpcServiceChain({
-    fetch: givenFetch,
-    btoa: givenBtoa,
-    serviceConfigurations: availableEndpointUrls.map((endpointUrl) => ({
+  const rpcService = new RpcServiceChain(
+    availableEndpointUrls.map((endpointUrl) => ({
+      ...getRpcServiceOptions(endpointUrl),
       endpointUrl,
     })),
+  );
+  rpcService.onBreak(({ endpointUrl, failoverEndpointUrl }) => {
+    messenger.publish('NetworkController:rpcEndpointUnavailable', {
+      chainId: configuration.chainId,
+      endpointUrl,
+      failoverEndpointUrl,
+    });
+  });
+  rpcService.onDegraded(({ endpointUrl }) => {
+    messenger.publish('NetworkController:rpcEndpointDegraded', {
+      endpointUrl,
+    });
+  });
+  rpcService.onRetry(({ endpointUrl, attempt }) => {
+    messenger.publish('NetworkController:rpcEndpointRequestRetried', {
+      endpointUrl,
+      attempt,
+    });
   });
 
   const rpcApiMiddleware =